+++ test/test_bridge.py
@@ -35,7 +35,7 @@ import pytest
 from distributed import Client, LocalCluster
 
 from deisa.dask import Bridge
-from utils import FakeComm, FakeCartComm
+from utils import FakeComm, FakeCartComm, async_map
 
 logging.basicConfig(level=logging.DEBUG)
 
@@ -45,10 +45,11 @@ class TestBridge:
     def env_setup(self):
         cluster = LocalCluster(n_workers=1, threads_per_worker=1, processes=True,
                                dashboard_address=":0", worker_dashboard_address=":0")
-        os.environ['DEISA_DASK_SCHEDULER_ADDRESS'] = cluster.scheduler_address
+        os.environ["DEISA_DASK_SCHEDULER_ADDRESS"] = cluster.scheduler_address
         client = Client(cluster)
         client.wait_for_workers(1, timeout=10)
         yield client, cluster
+        client.close()
         cluster.close()
 
     def get_new_bridge(self):
@@ -156,9 +157,13 @@ class TestBridge:
             }}
         comm_state = FakeComm.State(4)
 
-        bridges = [Bridge(comm=FakeCartComm(comm_state, rank, dims=(2, 2)),
+        def make_bridge(rank):
+            return Bridge(comm=FakeCartComm(comm_state, rank, dims=(2, 2)),
                           arrays_metadata=arrays_metadata,
-                          wait_for_go=False) for rank in range(4)]
+                          wait_for_go=False)
+
+        # Create bridges in parallel (Split is a collective op)
+        bridges = async_map(range(4), make_bridge)
 
         async def _bridge_send():
             await asyncio.gather(*[asyncio.to_thread(bridge.send, 'temperature',